Finalized pose translation, fixed int16_t limit bugs; expanded graphics capabilities to int32_t coordinates for out-of-bounds rendering necessitated by poses

+++ inc/graphics.hpp
@@ -53,11 +53,11 @@
 		void drawBonesReff(unsigned char currBone,const char *altCurrBoneParent);
 		
 		// Drawing Dynamic Model-State-Dependant UI
-		void drawCircle(int16_t x,int16_t y,bool isRadNorm,float radius,uint32_t color);
-		void drawCircleOutline(int16_t x,int16_t y,bool isRadNorm,float radius,uint32_t color);
+		void drawCircle(int32_t x,int32_t y,bool isRadNorm,float radius,uint32_t color);
+		void drawCircleOutline(int32_t x,int32_t y,bool isRadNorm,float radius,uint32_t color);
 		
-		void drawMark(unsigned int i,int16_t x,int16_t y,bool isScaleNorm,float scale);
-		void drawStem(unsigned int i,int16_t srcX,int16_t srcY,int16_t destX,int16_t destY);
+		void drawMark(unsigned int i,int32_t x,int32_t y,bool isScaleNorm,float scale);
+		void drawStem(unsigned int i,int32_t srcX,int32_t srcY,int32_t destX,int32_t destY);
 		
 		enum lineType{
 			LINE_VERTICAL,
